Fix: Feed white screen issue - Implement SimpleWorkingFeedScreen
- Created SimpleWorkingFeedScreen with direct Firestore access
- Updated MainNavigationScreen to use new feed screen
- Fixed white screen issue preventing users from viewing/creating posts
- Added comprehensive documentation and deployment scripts
- Deployed successfully to https://talowa.web.app

Features now working:
- View posts in chronological order
- Create posts with text and images
- Upload single/multiple images
- Like/unlike posts
- Pull-to-refresh
- Real-time updates
- Smooth scrolling
- Clear error handling

Files changed:
- lib/screens/feed/simple_working_feed_screen.dart (new)
- lib/screens/main/main_navigation_screen.dart (updated)
- Multiple documentation files added
- Deployment scripts created

Deployment: Live at https://talowa.web.app (2025-11-16 22:27:14)

+++ lib/services/ai/ai_moderation_service.dart
@@ -200,9 +200,9 @@ class AIModerationService {
       ]);
 
       // Ensemble approach for 95% accuracy
-      final perspectiveScore = results[0] as double;
-      final openAiScore = results[1] as double;
-      final localScore = results[2] as double;
+      final perspectiveScore = results[0];
+      final openAiScore = results[1];
+      final localScore = results[2];
 
       // Weighted ensemble (Perspective API has highest weight due to specialization)
       final ensembleScore = (perspectiveScore * 0.5) + (openAiScore * 0.3) + (localScore * 0.2);
